updates for using dao

+++ app/dao/UserDAO.scala
@@ -6,7 +6,7 @@ trait UserDAO {
     def findUser(userName: String, userPassword: String): User
     def findUser(userName: String): User
     def findUser(userId: Integer): User
-    def addUser(user: User): Long
+    def addUser(user: User): Integer
     def updateUser(user: User): User
-    def deleteUser(user: User): Int
+    def deleteUser(user: User): Integer
 }
